refactor: move primary button
