+++ packages/components/buttons/primary-button/src/primary-button.story.js
@@ -2,9 +2,9 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { withKnobs, boolean, text, select } from '@storybook/addon-knobs/react';
-import Section from '../../../../.storybook/decorators/section';
-import * as icons from '../../icons';
-import Readme from './README.md';
+import * as icons from '@commercetools-uikit/icons';
+import Section from '../../../../../.storybook/decorators/section';
+import Readme from '../README.md';
 import PrimaryButton from './primary-button';
 
 const iconNames = Object.keys(icons);
