+++ packages/components/buttons/primary-button/src/primary-button.visualroute.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { InformationIcon, PrimaryButton } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 export const routePath = '/primary-button';
