+++ src/components/fields/async-creatable-select-field/async-creatable-select-field.form.story.js
@@ -10,9 +10,9 @@ import {
   number,
 } from '@storybook/addon-knobs/react';
 import Spacings from '@commercetools-uikit/spacings';
+import PrimaryButton from '@commercetools-uikit/primary-button';
 import Section from '../../../../.storybook/decorators/section';
 import FormikBox from '../../../../.storybook/decorators/formik-box';
-import PrimaryButton from '../../buttons/primary-button';
 import SecondaryButton from '../../buttons/secondary-button';
 import Readme from './README.md';
 import AsyncCreatableSelectField from './async-creatable-select-field';
