+++ src/components/inputs/money-input/money-input.form.story.js
@@ -6,10 +6,10 @@ import { injectIntl } from 'react-intl';
 import { withKnobs, number } from '@storybook/addon-knobs/react';
 import omitEmpty from 'omit-empty';
 import Spacings from '@commercetools-uikit/spacings';
+import PrimaryButton from '@commercetools-uikit/primary-button';
 import ErrorMessage from '../../messages/error-message';
 import Section from '../../../../.storybook/decorators/section';
 import FormikBox from '../../../../.storybook/decorators/formik-box';
-import PrimaryButton from '../../buttons/primary-button';
 import SecondaryButton from '../../buttons/secondary-button';
 import Readme from './README.md';
 import MoneyInput from './money-input';
