+++ src/components/inputs/select-input/select-input.form.story.js
@@ -4,9 +4,9 @@ import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { withKnobs, boolean } from '@storybook/addon-knobs/react';
 import Spacings from '@commercetools-uikit/spacings';
+import PrimaryButton from '@commercetools-uikit/primary-button';
 import Section from '../../../../.storybook/decorators/section';
 import FormikBox from '../../../../.storybook/decorators/formik-box';
-import PrimaryButton from '../../buttons/primary-button';
 import ErrorMessage from '../../messages/error-message';
 import SecondaryButton from '../../buttons/secondary-button';
 import Readme from './README.md';
